move helper func from core to users
