+++ nautobot/core/models/utils.py
@@ -150,13 +150,6 @@ def serialize_object_v2(obj):
     return data
 
 
-def serialize_user_without_config_and_views(user):
-    serialized_data = serialize_object_v2(user)
-    for key in ["config_data", "default_saved_views"]:
-        serialized_data.pop(key, None)
-    return {"data": serialized_data}
-
-
 def find_models_with_matching_fields(app_models, field_names=None, field_attributes=None, additional_constraints=None):
     """
     Find all models that have fields with the specified names and satisfy the additional constraints,
