+++ nautobot/users/forms.py
@@ -9,8 +9,8 @@ from timezone_field import TimeZoneFormField
 from nautobot.core.events import publish_event
 from nautobot.core.forms import BootstrapMixin, DateTimePicker
 from nautobot.core.forms.widgets import StaticSelect2
-from nautobot.core.models.utils import serialize_user_without_config_and_views
 from nautobot.core.utils.config import get_settings_or_config
+from nautobot.users.utils import serialize_user_without_config_and_views
 
 from .models import Token
